Enhance code quality and documentation
- Added CI badge to README for build status visibility.
- Clarified Windows-specific limitations in README.
- Improved console output encoding handling in app.py using contextlib.
- Simplified fade_frames function in dsp.py.
- Enhanced type annotations in engine.py for better clarity.
- Refactored error handling in audio engine to use contextlib.
- Updated function signatures across various modules for consistency.
- Improved code readability by using pairwise in theme.py.
- Fixed minor issues in UI widgets for better performance and clarity.
- Introduced pyproject.toml for better dependency management and project metadata.
- Updated requirements.txt to align with pyproject.toml.
- Refined tests for better coverage and clarity.

+++ mp3player/core/audio/dsp.py
@@ -17,7 +17,7 @@ FADE_MS = 10.0
 
 def fade_frames(sample_rate: int, ms: float = FADE_MS) -> int:
     """How many frames a full 0 -> 1 fade spans at `sample_rate`."""
-    return max(1, int(round(sample_rate * ms / 1000.0)))
+    return max(1, round(sample_rate * ms / 1000.0))
 
 
 def resample(
